feat: Implement entering the game
The world now loads, and the client is fully configured when entering
the "play" state. This includes loading and replaying the registry
packet which I recorded from a "real" server via Wireshark and stored
as a sequence of hexadecimal records so that COBOL can load it.
Many additional client-bound packets are also now implemented,
except for the world chunk packet, which is difficult to generate and
will be done later.
The keep-alive mechanism is very hacky (sending a keep-alive packet
whenever a packet is received from the client), but it works for now
to keep the client connected.