feat: Begin implementation of a timed game loop (#19)
Before this patch, receiving packets from the client was a blocking
operation, and it was impossible to do basically anything else at all.
Now, each socket read only gets a very short timeout, so no matter
how much data could be read, the call will complete - and we will be
able to perform other stuff before trying to read again.

Also in this patch, a custom C++ util module is added that has one
function implemented to retrieve the system time in milliseconds.
Using this, the game update + socket read loop can be timed
precisely enough to achieve Minecraft's 20 ticks per second. The idea
is to first perform any game logic, then spend the rest of the tick
doing socket communication.

Since the read is now (somewhat) async, it should be possible to
extend this for multiplayer as well (later).