feat: Send chunk data
When logging in, we now send a 9x9 area of chunks to the client.
The chunks consist of stone up to a certain height, with air above.
Lighting will have to be fixed in a follow-up patch - for now, everything
is dark. But still, the world now has stuff in it!

Fixed encoding of negative integers as a drive-by.