feat: Implement Encode-Double for player position (#31)
This adds a C++ utility function for converting an IEEE 754
double-precision floating-point number to a byte array. This is needed
to encode the player position, for example. I thought about doing this
purely in COBOL but even if I spent enough time to ensure a bug-free
implementation, the performance would likely be horrible. This way it's
fast and accurate with very little code.