feat: Store chunk data in RAM (#33)
When the server is starting, it now generates initial world data that is
stored in RAM. When a player logs in, the chunks are encoded
dynamically and sent to the player. This is half of what is needed for
block placement/destruction to be effective - the other half is actually
handling the place/destroy packets sent by the client and updating the
data in RAM.

For now, a fixed 7x7 chunks are generated, since the performance is not
yet optimized for more, and the network usage is quite high.