feat: Support multiple connections (but not players) (#39)
This patch adds support for responding to ping/status requests even
while someone is playing. This was previously not possible, as only a
single client socket could be open at any given time, and no other
sockets would be accepted until the client disconnects.

Now, the storage in RAM is reserved for up to 10 clients, each with
their own state, package buffers, keepalive timers, and so on.

However, there can still only be a single player. When another client
wants to enter the play state, it will be disconnected with a "Server is
full" message. Accordingly, the status response packet is updated to
indicate the current player count for clients to show before joining.