feat: Extract registries data from server.jar (#58)
To allow players to place arbitrary blocks and use any of the items the
game has to offer, we need access to a list of these things. Fortunately,
this data is available from the official Minecraft server .jar file via an
integrated command-line interface.

This patch adds a Makefile target that downloads the .jar and
extracts relevant data from it, resulting in multiple JSON files.
Furthermore, this patch adds a handmade JSON parser that is used
to parse the "registries.json" file at run-time. The JSON parser is
covered by unit tests.

The registries data is not yet widely used, but as a proof-of-concept,
the protocol ID of the player entity type is now extracted from it
instead of being hardcoded. This, along with future patches, will
greatly reduce the amount of work needed to support additional
features and perform version upgrades.

Finally, this patch overhauls the Dockerfile such that the data
extraction can be performed at build time, while keeping a slim final
image with only the relevant parts.