feat: Load block state IDs from blocks.json
Continuation of #58. Similar to that PR, the protocol IDs of block states
are now loaded from the blocks.json data file at runtime. This improves
robustness for future changes. It also prepares us to correlate item IDs
to block IDs in a future patch, to enable placing arbitrary blocks.

To achieve this goal, the JSON parser is extended to support parsing
arrays. The blocks.json file is parsed once at startup and its contents
stored in a shared data structure such that the default block state ID of
any named block can be easily retrieved.