feat: Support all the blocks
Building upon PRs #58 and #62, players are now able to place any of
Minecraft's blocks in the world! (With some caveats.) This works by
obtaining the item name from the player's current item ID, then
checking for a block with the same name as that item. Caveats include
things such as water, which doesn't use the corresponding item but
rather a separate water bucket item, and being limited to the default
block state. The latter causes glitches with blocks such as stairs, which
are orientated incorrectly. Still, this is a major improvement!