feat: Use UUIDs to differentiate players
This patch implements somewhat proper handling of player UUIDs.
Instead of using the index in the PLAYERS array as the UUID, which
would break as soon as player data can be offloaded to disk, a
persistent UUID is computed from the player name on login and
subsequently used in all packets.

Previously, it was possible for the same player (same username and
same UUID) to log in multiple times to the server for testing purposes.
This is no longer possible with this change. However, the same UUID can
still be used as long as the username is different, just like on a regular
Java offline-mode server.