fix: Send all blocks as fully illuminated
The assumption in #27 was incorrect. The client actually doesn't
calculate the light masks correctly when joining the server, at least not
in all circumstances. This patch is a simple fix that sends all blocks as
fully lit, which is certainly better than random dark patches. However,
for an optimal experience, we would need to compute the lighting
properly using a flood-fill algorithm, but that is outside the current
scope of this project.