fix: Disconnect existing player when connecting twice (#77)
When a player logs in with a UUID that is already linked to a player on
the server, this existing player is now kicked and the player logging in
is allowed to do so. This is in contrast to before this patch, where a
player could not connect if a previous connection was still live, which
can cause issues in case of network timeout, for example.