feat: Implement saving and loading of world data (#84)
When the world is first generated, its data (world age and time of day)
as well as the chunk contents are saved to disk. This data is also saved
whenever the "stop" or "save" commands are used - the latter being
a new addition in this patch. On server start, the 7x7 chunk area
centered at (0, 0) is loaded from disk. Provisions are made such that
further chunks can be dynamically saved and loaded in future patches.

The file format for chunk data is similar to the network packet, in that
it is split into 16x16x16 sections (= 24 sections per chunk). Each section
has a block palette whose length depends on the number of distinct
blocks used within the section, and an array of 4096 (16x16) indices
into that palette, representing the individual blocks.

Similarly to Minecraft's own "anvil" chunk format, the block palette
stores (for each entry) the block name such as "minecraft:bamboo_stairs"
and the block state properties such as "facing": "north" or "half": "top".
Since CobolCraft uses numeric block state IDs internally to simplify the
sending of network packets and reduce memory usage, new conversion
methods were implemented between block state IDs and the
representation based on name + key/value properties.

On the file system, each chunk gets its own file. This is known to be
suboptimal and was one of the downsides of the "region" file format
used by Minecraft itself before the switch to "anvil". However, this is
much simpler than grouping chunks into fewer files and probably not
the limiting factor in CobolCraft's performance.