feat: Save and load player data (#88)
With this patch, player data is saved and restored from disk, meaning
that persistence in general is complete. The player is saved when the
world is saved, or when the player disconnects. Loading happens when
the player connects again. As a side-effect, this means CobolCraft can
now support a very large (basically infinite) amount of offline players,
which was previously not the case.