fix: Improve efficiency and correctness of block states (#92)
This patch gets rid of the block states array, which consumed a lot of
memory, in favor of storing the minimum/maximum IDs only (since
block state IDs are contiguous for a given block). Additionally, a bug is
fixed where the final block state would be assumed as the default state
instead of using the actual default.