feat: Implement infinite world generation and load/unload
The server now keeps track of the center of each player's view area,
i.e., the chunks visible to the player. Whenever this area changes, the
chunks coming into range are enqueued for sending. One chunk per
tick is sent for any given player. Chunks are loaded from disk or newly
generated, if needed. Additionally, chunks outside the view area of
any player are automatically unloaded to free memory.

Before this patch, the world would only ever have 49 chunks, and those
would be sent to the player once on login. This caused a relatively
large amount of lag for the entire server since chunk encoding is not
yet optimized. Now, since we send chunks quite often on player
movement, the lag is not restricted to login, but happens frequently.
It will be necessary to improve the performance of chunk encoding to
remedy this. However, I think it's pretty awesome that the world is now
infinite, and am sure it will work smoothly a few more patches down
the line. :)