fix: Avoid storing/transmitting empty chunk sections
We now keep track of the number of non-air blocks in a given chunk
section. Sections that are fully empty (only consisting of air) are neither
stored on disk nor sent over the network to players. On my machine,
this reduced the stored size per chunk from ca. 250 kiB to ca. 80 kiB
and reduced the time to transfer a single chunk to a client from 53ms
to 25ms. Notably, 53ms is longer than the time available for a game
tick, while 25ms is less than a game tick, meaning the game is again
fast enough to actually play!