fix: Optimize parsing of blocks.json
This patch improves the performance of JSON parsing, and parsing of
blocks.json in particular. Overall, a reduction from 1050ms down to
460ms is observed.

The benefits come mostly from the removal of redundant operations:
Explicit failure checks, for instance, can often be avoided without
downsides if the next instruction(s) are able to deal with a failure
condition themselves already.

Further improvements are possible by avoiding string parsing in
JsonParse-SkipValue, and avoiding repetitive calls to LENGTH.