feat: Add support for stone stab placement (#107)
The block "minecraft:stone_slab" can now be placed in the top, bottom,
and double configurations. All scenarios regarding placement next to,
on top of, or below existing slabs are handled.

Minecraft has dozens of slab types overall, not just stone. They can be
detected via the `definition.type` property in blocks.json, which is set to
"minecraft:slab". This is a possible way to register the handler for all
slab types easily - this will be done in a follow-up patch.