refactor: Make some packet's ID dependent on client state
Some packet types are used in more than one client state. This is the
exception, as most packets are used in a single state only. However,
since these packets are entirely identical except for their ID when used
in different states, it makes sense to have just a single file per packet
type.