feat: Send disconnect message to existing player on re-login
Some packets are used throughout different client states, which this
patch implements for both packets where this is currently relevant:
clientbound plugin message, and client disconnect. The latter
unfortunately uses different encoding between the states (JSON vs
NBT text components), but since the usage is identical, this is still
included in a single packet routine.

In doing so, we can get rid of a TODO in server.cob: When a player
joins the server while already being connected on a different TCP
connection, the existing connection now receives an explicit disconnect
message before being cut off.