refactor: Move command handling to separate file (#118)
This patch moves a lot of code around with the goal of extracting
console command handling from server.cob into a separate file to aid
with readability. This necessitates a larger refactor of the server code to
also extract chat broadcasting, as well as multiple server-related
features, into their own programs.