fix: Store packet sending error until next tick (#120)
This patch adds a field to the client definition to store a possible
sending error. If this field is populated, no further writes will be made
to the socket, and the client will be disconnected on the next tick. It is
automatically populated by the "SendPacket" program.

There are two large advantages to this method of error handling:
- If packet sending is part of a larger algorithm, such as block
    placement, there no longer need to be any precautions in place to
    protect the algorithm as a whole from failing when a packet fails.
- Sending packets becomes much simpler overall, no longer
    requiring a client handle or a path to propagate errors, but only a
    widely-available client ID. This makes it easy to integrate into parts
    of the codebase such as "item use" handlers, where it would have
    previously caused too much complexity.