feat: Send available commands list to the client (#127)
Previously, the client would show any command input as invalid, since
it didn't know about the available commands (although it could run
them just fine if the user ignored the warning). Now, we send a data
structure containing all available commands and their arguments to
the client when connecting. The structure is a (potentially cyclic, but
usually tree-like) graph of literals and arguments.

For now, creating this structure has to be done with lots of manual
effort, but this can be further automated in the future when
commands are registered dynamically.