feat: Add NBT encoding and decoding capabilities
This patch adds subroutines for working with NBT, along with tests.

Encoding is rather straight-forward: The caller invokes, in sequence, the
methods that will produce the desired NBT data.

Decoding is somewhat more involved as it requires us to maintain a
stack of data structures preceding the current value, that is, the
sequence of lists and compound tags preceding the current tag. That
is due to the fact that tags may have a name after their type ID, but
only if they are inside a compound tag.

This is not an exhaustive implementation of all tag types, but it is
enough to save a `level.dat` file according to our needs. Rewriting the
world save-file to NBT is planned as an upcoming patch.