refactor: Use state for NBT encoder
Like we already do for the NBT de-coder, we now keep track of the
state of NBT encoding, such as the stack of compound/list tags up to
the current level. For now, the only benefit is additional correctness
checks, but once encoding of list tags is added, we will use it more
concretely.

Furthermore, this patch combines the "Encode-NamedXYZ" and
"Encode-XYZ" subroutines into a single one with optional arguments
for the name and name length, to reduce the amount of
semi-duplicate code needed.