feat!: Save player data in NBT format (#141)
Similarly to #133, this patch migrates another data file away from plain
COBOL records and towards Minecraft's official NBT format: The player
save data. We continue storing the same information as previously.
Property names were chosen to be identical to the ones Minecraft uses
internally, and by opening the save data in any of the available NBT
editors, we can see that the generated binary output is indeed valid
NBT! :)

As a caveat, we don't yet convert the items' NBT data tags to NBT
compounds on disk, but store this data as a byte array instead for
simplicity. However, this should be changed in the future.