feat: Add a utility to count leading zero bits
This patch adds a C++ utility function for efficiently counting the
number of leading 0 bits in a 32-bit unsigned integer. This will be
needed for computing the stride (number of bits per entry) in
paletted data structures such as the NBT chunk format.

Splitting it out of the main patch so it's easier to digest for anyone
reading this.