feat!: Save chunk data in NBT format
This is the final step towards saving all data in a format compatible
with the official Minecraft server, after #133 and #141. Now, the chunk
data is also saved as NBT, including a full palette implementation. This
latter feature causes a very nice reduction in file size for chunks with
few distinct blocks compared to the previous format based on native
COBOL records. Encoding and decoding is of course much more
involved, however, the performance is still good, and there are many
advantages regarding extensibility of the format (and perhaps, at
some point, we can load worlds saved by Minecraft itself?).