feat: Use offset-based encoding (#152)
While the JSON and NBT encoders already used an "offset" parameter
for locating into the output string, the primitive encoding methods
(number-to-byte conversion, VarInt, and some Minecraft-specific types)
did not, and instead always wrote into the output buffer starting from
the beginning. This is rarely what is needed, especially when building
packet payload, since there will be multiple sequential encoding calls.
Hence, it was almost always necessary to use a temporary buffer into
which the data is encoded, before being moved into place. This is not
only less performant, but also more difficult to reason about due to
the larger amount of code needed.

This patch converts all encoding functions to take an offset into the
buffer (by reference), which they "automatically" increment before
returning. This means that we can remove temporary buffers from
almost everywhere and simply chain encoder calls without manually
incrementing the offset.

However, there is a disadvantage in that we need to subtract 1 to get
the payload length of packets before sending. The `COMPUTE`
statement that does this is repeated once per packet. I think this is
acceptable given the significant boilerplate reductions overall.