fix: Decode packet length only once all bytes are read
The ReceivePacket logic was flawed such that longer packets, where
the length VarInt consists of two or more bytes, would incorrectly try
to decode their length after just a single byte was received. This
caused a failure, and the client would be disconnected.

Note that very few server-bound packets can even be long enough
to trigger the bug, which is why it was undetected for so long.