fix: Implement float/double coding without C++
Since GnuCOBOL represents variables of type FLOAT-SHORT and
FLOAT-LONG in memory as IEEE-754, there is really no need to use
C++ to perform the float-to-byte conversion. Instead, we can directly
re-interpret the data as PIC X and use REVERSE to obtain a big-endian
representation.

Not only is this the nicer way of doing it, but it leads to roughly 15%
performance gains for encode/decode microbenchmarks.