fix: Implement C++ utils as individual functions
Before this patch, there was a single C++ function COBOLCRAFT_UTIL
that multiplexed the call based on a two-digit "op-code". This design
was chosen because it mirrored the way the CBL_GC_SOCKET library
works. However, not only is this harder to maintain and less obvious
since one needs to know the meaning of each op-code, but it is also
less performant due to the string operations required.

Now, there is one C++ function for each possible operation. The
parameters are also declared such that the proper COBOL argument
type can be directly passed in, removing the need for the separate
util.cob wrapper.