fix: Use reverse function for encoding and decoding
We previously did not use FUNCTION REVERSE in encoding and
decoding because GnuCOBOL wouldn't compile when using reference
modification on the buffer with a literal length, due to the buffer being
defined as ANY LENGTH. However, when a variable length is used, this
becomes possible.

This patch leads to performance improvements across the board.
Specifically, coding of 64-bit integers is now 40% faster; for 32-bit
integers, the improvement is 15%, and for 16-bit integers, 4%.

As a result, operations such as saving/loading world files or sending
chunk data over the network should be noticeably faster as well due
to their usage of large 64-bit integer arrays.