feat: Implement gzip compression for level and player data
Minecraft saves NBT files, such as level.dat or the player `<uuid>.dat`
files, in a gzip-compressed format. With this patch, CobolCraft now
does the same, meaning that we are able to load world metadata and
player data from existing world saves! More work is however needed
regarding the chunk format.

This adds a dependency on zlib, which must be available on the
system during compilation (headers) and at run time (shared library).