fix: Make chunk NBT parsing order-independent (#170)
While we write our chunk NBT data in the order that makes the most
sense for efficient parsing, Minecraft itself emits NBT tags in a way
that appears almost random (perhaps a HashMap?). If we want to
succeed in loading existing chunk data,  we need to deal with this (at
the cost of some performance).