feat!: Implement "Anvil" / "Region" file format (#171)
CobolCraft now reads and writes chunk data in the same format used
by Minecraft itself. Chunks are no longer stored in individual files, but
in region files containing up to 32x32 (= 1024) chunks each. There is
full support for all modes of compression (gzip, deflate, and
uncompressed) - CobolCraft always uses Minecraft's preferred mode
of compression (deflate), but is able to read any other mode as well.
In other words: CobolCraft should now be able to load existing
worlds!

Region files are divided into sectors of 4096 kiB, and each chunk
takes up one or more such sectors, unless it has not been generated
yet. Most chunks will consume just one or two sectors (especially if
compressed), however, much larger chunks are possible, and chunk
size may change over time depending on the chunk contents. Hence,
writing to region files is an identical problem to memory page
allocation as performed by operating systems. The start sector and
sector count of each chunk is stored in a table inside the region file
header. From this table, we can compute a sector usage array, where
the boolean at position N is set if (and only if) the N-th sector is
currently used. We allocate sectors for a chunk-to-be-saved via the
first-fit strategy by looking for a "run" of the required number of free
sectors within the usage array.

Note that the chunk NBT format stored within the region format is
not changed as part of this patch because it was already fully
compatible to the official format.