fix: Optimize chunk loading
This patch significantly reduces the chunk loading time in a flat world,
and likely in any other world type, too. Performance improvements of
55% were measured. The largest reductions come from using an
INITIALIZE statement to set an entire chunk to a single block type, and
from precomputing `2 ** PALETTE-BITS` and using a single DIVIDE
statement with a REMAINDER clause.

Each partial change in this patch was independently checked to
ensure there are no regressions anywhere.