fix: Inline Encode-UnsignedLong in chunk data packet
For the chunk data packet, the most time is spent encoding the block
state palette. We can achieve a 20% performance gain by inlining the
operations of Encode-UnsignedLong directly into the packet code.
While this breaks the "separation of concerns" principle, this packet is
quite heavy overall, so this seems worth it.

Note that adding the `STATIC` keyword to the call yields no
improvements. The call has to be avoided altogether.