feat: Add support for block entities (and add beds)
This patch started out with just adding placement of beds, which are
interesting because they consist of two individual blocks that need to
be set simultaneously when placed. However, I quickly noticed that
beds require support for block entities, or they will become invisible
to players (since the block entities are used for rendering). As such,
this patch now adds block entity support along with the bed blocks,
which now serve as a demo block for the new feature.

For simplicity reasons, block entities are currently stored within each
chunk in an array large enough to hold one entity per block. This is
not the most memory-efficient representation but avoids dynamic
memory allocation - this may be needed in the future, depending on
performance measurements.

As usual, the block entities are saved within the region files in the
same format used by Minecraft itself, ensuring that CobolCraft is able
to load pre-existing worlds.

The chunk data packet is updated to send block entities. Further, a new
packet is added to send a single new block entity upon creation.

Finally, the World subroutines have been modified such that the bed
item callback is able to create block entities upon placement, and also
to ensure that block entities are removed when the respective block is
removed or replaced. An internal block entity counter is kept per chunk
to short-circuit block entity handling in the common case of zero
block entities.