feat: Add block destroy callbacks to handle two-part blocks
Blocks can now register special logic to execute when they are broken,
overriding a generic handler (that simply sets the broken block to air).
Such callbacks are added for door and bed blocks, the only two-part
blocks currently implemented, to ensure that both parts are destroyed
when one is broken by the player.