feat: Implement solid/non-solid block faces (#188)
Some blocks require a solid block surface next to them. For instance,
torches cannot be placed on non-solid blocks (with some exceptions).
To make this work, a new type of block callback is added via this patch
to check whether a block is solid at the given facing direction. This is
then used to implement a check for torches to ensure they only
attach on such faces. A generic variant of the callback exists as a
fallback for all blocks unless a specialized handler is registered.

This patch adds special handlers for the following block types:

* air (no solid faces)
* beds (no solid faces)
* doors (no solid faces)
* slabs (solid faces depending on top/bottom/double slab)
* torches (no solid faces)
* trapdoors (solid at the top or bottom, depending on their state)