feat: Send "update tags" packet
The "update tags" packet is required to inform the client about some
game aspects. Most important for us is the `minecraft:fluid` property
which specifies the supported fluid types (water, lava). If this is not
sent, the client is unable to render lava and renders it as water
instead.

Due to the large amount of content within the packet, it is sent as
a blob for now, however, we may construct it manually at a later
point if the need arises to customize some aspects of it.