feat: Add support for buckets (water, lava, empty)
This patch adds support for the main bucket types, allowing players
to place and remove water and lava source blocks. This is rather
straight-forward for the most part, as long as the player is clicking on
a non-fluid block to place or remove a fluid next to it. However, the
player may also be far enough away such that they are not in reach
of any such block, but would be in reach of an existing fluid block.
This scenario does NOT trigger a "use item on" packet to be sent.
Instead, the client sends a "use item" packet as if the player was
clicking in air, while still expecting the server to calculate the liquid
block as the click target and act on it. This requires ray-tracing to be
implemented on the server, which is not on my agenda, yet. Hence,
this patch simply acknowledges the "use item" action without
updating any blocks, thereby preventing direct interactions with fluid
blocks to avoid desynchronization issues.