feat: Add support for block replacement
When the player right-clicks with a block in hand, the position where
this block should be placed is subject to a condition: If the block
currently at the click position is considered "replaceable", the new
block should be placed right there. Otherwise, it should be placed next
to the clicked position (depending on which side of the block was
clicked).

Before this patch, the code didn't take this into account - always
placing the new block adjacent to the clicked position. Now, this is
handled for the following replaceable block types:

- air (all three kinds)
- liquids (water, lava)
- tall grass

There may be further kinds of replaceable blocks in the game; these
can be added in the future via an extendable callback mechanism.