fix: Obtain dimension_type and chat_type from registry (#199)
Continuation of #195 to make use of the data that is now available
for chat and dimension types.

Note: For some reason, the chat type needs 1 added to it. This was
already the case when using the blob, however. It seems to have
changed between 1.20.6 and 1.21. Should keep an eye on this packet.