feat: Add a persistent JSON file-based whitelist
The whitelist can now support up to 100 players. It is stored in a
whitelist.json file in the server directory, using the same format as the
official server to be drop-in compatible.

Console and in-game commands are also added to reload the whitelist,
list whitelisted players, and add/remove players.