refactor: Split commands into files and make registerable (#208)
The server/in-game commands are no longer part of one big source file,
but can instead be registered similar to block and item callbacks, using a
program pointer for execution. One limitation still exists in that the node
structure (for client-side auto completion) must still be hardcoded. This
will be addressed in a future patch.

To enable this change, the `HandleCommand-SendToClient` subroutine is
moved to chat.cob as `SendChatMessage` and used throughout.