refactor: Make command tree setup part of registration (#213)
This shouldn't impact the behavior at all, but now, we no longer hardcode
the command tree globally, but instead define each command's nodes
in its registration program. This greatly improves the command-related code
structure. In the future, we could also generate the usage text of each
command automatically based on this information.