feat: Allocate client receive buffers dynamically
See #215. Instead of declaring the buffer statically for each client, which
means the maximum amount of memory will always be consumed, it is
now a pointer and is allocated/freed when a client connects/disconnects.
In effect, unused client slots consume very little memory.

To actually access the buffer, it must be moved to a 01-level BASED item.
BASED items cannot be at a level other than 01 or 77, hence the pointer
indirection.