refactor: Remove obsolete socket wrapper programs
Follow-up to #230. Because we're now using custom C++ network functions
that fit the specific needs of CobolCraft, it is possible to remove the
COBOL wrapper programs, reducing complexity and improving performance.

One slight change had to be made to `SocketWrite()` to handle buffers
larger than 64 kB, which was previously done by the wrapper.