fix: Receive entire packet at once if possible
Another follow-up to #230. Since reading data from a socket is no longer
associated with a 1ms delay in the worst case, we can attempt to read the
entire packet (length + ID + payload) in one go. This eliminates the need
to do more than one SocketPoll. For short packets (1-byte length and
no payload), we previously had to do 2 polls, and in the worst case (5-byte
length and 2 MiB payload) there could be 37 polls.

Overall, receive performance should be greatly improved, which can also
be seen in Valgrind profiles, where total time spent in SocketPoll is halved.