fix: Always set components for loaded inventory items
When loading a player.dat file generated by the official server or client, it
will not have "tag" entries for the inventory slots like it did in the past, but
recent versions instead use a "components" compound.

At this time, CobolCraft is unequipped to properly deal with this. Yet, we
can cheat a bit and default to the empty network representation of item
components. This allows the player to join, at least, even if enchantments
etc. will be removed from items.

Inventories saved by CobolCraft should be unaffected by this change.