feat: Add support for biome data in chunk sections
This implements encoding/decoding of biome data when saving/loading
to/from disk, as well as (crude) encoding of biome data in the chunk data
packet. Finally, we can have more than just the plains biome.

For all encoding of biome data, Minecraft uses the same "paletted
container" structure as for blocks, but the details are different enough that
the code is not easily generalized (at least not in COBOL).

As a refactor, the SkipUntilTag and SkipRemainingTags helper programs are
moved to nbt-decode.cob such that the chunk loading can be split into
different parts that are all able to call the aforementioned helpers, which is
not possible for sibling programs.