refactor: Implement palettes generically for chunk data packet
The previous implementation of paletted containers for block states and
biomes relied on hard-coded bits-per-entry values, which are actually
dependent on the number of potential unique IDs (i.e., the number of
registered block states, or the number of entries in the biome registry).
Hence, the values may change if a future Minecraft update adds enough
entries to require additional bits.

With this patch, there is now a single, generic sub-program for palette
encoding. This program is used for both block states and biomes, and the
bits-per-entry value is computed from the actual registry information.
This makes the code future-proof, and also much nicer to look at.