fix: Make game tick timing more precise
Due to the way we perform game ticks, each tick is very slightly too long.
The idea with this patch is to improve (1) the measurement precision and
(2) the idle sleep duration, thereby reducing the average overshoot.

Tests show a 10x improvement, from 0.5ms down to 0.05ms on average.

Keepalive and autosave timers were updated accordingly.