feat: Implement debug samples
By pressing F3+2, the client can display a graph of the server tick
performance over time, including the time taken by the main server logic,
secondary tasks, and idle time. This patch adds debug sample collection
and client subscriptions such that we can see a rough estimate of the server
performance in-game, which is useful because it lets us see immediately
how the server reacts to in-game actions.