feat: Use packet IDs from generated packets.json (#273)
We previously hard-coded all the packet IDs, which requires significant
amounts of manual work when updating to a new Minecraft version. This
patch introduces a parser for `data/generated/reports/packets.json`,
allowing us to go from packet reference to numerical ID and vice versa.

For serverbound packets, we can convert from ID to packet name, and use
that in the `EVALUATE` statements. For clientbound packets, there is a
new set of copybooks for initializing the numerical packet ID once from
the packet reference (client state + direction (clientbound) + packet).